// src/app/models/task.ts
import {Subtask} from './subtask';

export type Task = {
  id: string;
  title: string;
  description: string;
  subtasksIdsSequence: string[];
  subtasks: {[key in string]: Subtask};
};

export type CreateTaskData = {
  boardId: string;
  boardStatusId: string;
  title: string;
  description: string;
  subtasks: string[];
};

export type CreateTaskResult = {
  id: string;
  tasksIdsSequence: string[];
  subtasksIdsSequence: string[];
};

export type UpdateTaskData = {
  id: string;
  boardId: string;
  status: {
    id: string,
    newId?: string
  },
  title: string;
  description: string;
  subtasks: {
    id?: string;
    title: string;
  }[];
};

export type UpdateTaskResult = {
  id: string;
  subtasksIdsSequence: string[];
};

export type DeleteTaskData = {
  id: string;
  boardId: string;
  boardStatusId: string;
};

export type DeleteTaskResult = undefined;